PPHA-475: Ensure a response set is created on submission
Before all questionnaire pages unsure a response set exists for the
user or redirect back to start if they have submitted responses in the
last year.

This replaces the /start page submission and creation of Participant
and corresponding response set.

Using NHS Login flow, no post can be made to create the response set so
handle it on each POST request using the EnsureResponseSet mixin.