PPHA-554: Create infra-as-code for Spoke resources
This is an configuration and subtle upgrade to the Bicep code to deploy the foundation in spoke subscription for Terraform to take over the rest of the deployment. This will update:

The Container App Environment CAE.
The Container App
The PostgresDB
The Private End Points

It will also include the instructions for building a new environment.

Please note that the Dev environment has been deployed and is currently in operation.